+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-All notable changes to the "Terminal Sound Alerts" extension will be documented in this file.
+All notable changes to the "TerminalTone" extension will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.0] - 2024-01-01
 
 ### Added
-- Initial release of Terminal Sound Alerts
+- Initial release of TerminalTone
 - Play success sound when terminal command exits with code 0
 - Play failure sound when terminal command exits with non-zero code
 - Play warning sound when command succeeds but output contains warning keywords